refactor dashboard view model and related models and extensions
