+++ SunsetWorkout/Models/QuoteTag.swift
@@ -1,0 +1,10 @@
+//
+//  QuoteTag.swift
+//  SunsetWorkout
+//
+//  Created by Paul Oggero on 06/10/2022.
+//
+
+struct QuoteTag: Codable {
+    let key: String
+}
